hobby page - turned blog tracker related pages into components, added some comments
